fix camera issues

+++ src/camera.c
@@ -18,21 +18,21 @@ viewport_layout_t viewport_configs[4][4] = {
     { {0, 0, 160, 120}, {160, 0, 160, 120}, {0, 120, 160, 120}, {160, 120, 160, 120} }
 };
 
-void camera_init(camera_t *cam, int world_width, int world_height, int screen_width, int screen_height, float spawn_x, float spawn_y) {
+void camera_init(camera_t *cam, float world_width, float world_height, float screen_width, float screen_height, float spawn_x, float spawn_y) {
     cam->width = screen_width;
     cam->height = screen_height;
     cam->world_width = world_width;
     cam->world_height = world_height;
 
-    // --- CENTERING CALCULATIONS ---
-    // Subtract half the screen dimensions from the player spawn coordinates
-    int desired_x = (int)spawn_x - (screen_width / 2);
-    int desired_y = (int)spawn_y - (screen_height / 2);
+    // --- CENTERING CALCULATIONS (Using Floats) ---
+    // Subtract half the screen dimensions from the player spawn coordinates safely
+    float desired_x = spawn_x - (screen_width / 2.0f);
+    float desired_y = spawn_y - (screen_height / 2.0f);
 
     // --- MAP BOUNDARY CLAMPING ---
     // Prevent the camera from scrolling past the left/top edges
-    if (desired_x < 0) desired_x = 0;
-    if (desired_y < 0) desired_y = 0;
+    if (desired_x < 0.0f) desired_x = 0.0f;
+    if (desired_y < 0.0f) desired_y = 0.0f;
 
     // Prevent the camera from scrolling past the right/bottom edges
     if (desired_x + cam->width > cam->world_width) {
@@ -42,47 +42,41 @@ void camera_init(camera_t *cam, int world_width, int world_height, int screen_wi
         desired_y = cam->world_height - cam->height;
     }
 
-    // Set the finalized, safely-bounded starting positions
+    // Set the finalized, safely-bounded starting positions as floats
     cam->x = desired_x;
     cam->y = desired_y;
 }
 
-void camera_update_split(camera_t *cam, int p_x, int p_y, int p_w, int p_h, int view_w, int view_h, facing_dir direction, float dt) {
+
+void camera_update_split(camera_t *cam, float p_x, float p_y, float p_w, float p_h, float view_w, float view_h, facing_dir direction, float dt) {
     cam->width = view_w;
     cam->height = view_h;
 
     // 1. Calculate the standard dead-center baseline focal point
-    int desired_x = p_x + (p_w / 2) - (view_w / 2);
-    int desired_y = p_y + (p_h / 2) - (view_h / 2);
+    float desired_x = p_x + (p_w / 2.0f) - (view_w / 2.0f);
+    float desired_y = p_y + (p_h / 2.0f) - (view_h / 2.0f);
 
-    // =========================================================================
-    // --- THE VISUAL EXTENSION FIX: FORWARD-FACING FOCUS OFFSET ---
-    // =========================================================================
-    // Look 48 pixels ahead of the player's current trajectory path line.
-    // If your viewports are small (like the 160px wide quad grid), we scale 
-    // the offset down to 24 pixels so the camera doesn't jump too wildly.
-    float dynamic_bias = (view_w <= 160) ? 24.0f : 48.0f;
+    // Forward-facing focus offset
+    float dynamic_bias = (view_w <= 160.0f) ? 24.0f : 48.0f;
 
     if (direction == FACING_LEFT) {
-        desired_x -= (int)dynamic_bias; // Shift camera left to show what's ahead on the left
+        desired_x -= dynamic_bias;
     } else if (direction == FACING_RIGHT) {
-        desired_x += (int)dynamic_bias; // Shift camera right to show what's ahead on the right
+        desired_x += dynamic_bias;
     }
 
-    // 2. Enforce absolute level boundaries safely
-    if (desired_x < 0) desired_x = 0;
+    // 2. Enforce absolute level boundaries safely without integer truncations
+    if (desired_x < 0.0f) desired_x = 0.0f;
     if (desired_x + view_w > cam->world_width)  desired_x = cam->world_width - view_w;
-    if (desired_y < 0) desired_y = 0;
+    if (desired_y < 0.0f) desired_y = 0.0f;
     if (desired_y + view_h > cam->world_height) desired_y = cam->world_height - view_h;
 
-    // 3. Smoothly glide this camera toward its target
-    int error_x = desired_x - cam->x;
-    int error_y = desired_y - cam->y;
+    // 3. Smoothly glide using pure floating-point math
+    float error_x = desired_x - cam->x;
+    float error_y = desired_y - cam->y;
     
-    // We drop the tracking speed slightly (from 6.0 to 4.5) to give the camera 
-    // a beautifully organic trailing "ease-in" effect when changing directions.
     float tracking_speed = 4.5f;
 
-    cam->x += (int)((float)error_x * tracking_speed * dt);
-    cam->y += (int)((float)error_y * tracking_speed * dt);
+    cam->x += error_x * tracking_speed * dt;
+    cam->y += error_y * tracking_speed * dt;
 }
